rename FocusDetector js file
