+++ src/components/common/focus-detector.js
